+++ app/services/web_client/action_handler.rb
@@ -4,14 +4,15 @@ class WebClient::ActionHandler
   attr_reader :nkey, :caller, :notifications, :notice
   validate :nkey_should_be_correct
 
-  def initialize(nkey, caller: :web_client)
+  def initialize(nkey, owner: nil, caller: :web_client)
     @nkey = nkey
+    @owner = owner
     @caller = caller
     @notifications = Notification.by_nkey(nkey)
   end
 
   def owner
-    notification.contact.owner
+    @owner || notification.contact.owner
   end
 
   def notification
@@ -69,10 +70,10 @@ class WebClient::ActionHandler
   end
 
   def subscribe_or_unsubscribe(action_method, new_status)
-    if notification.contact.owner.send("#{new_status}?")
+    if owner.send("#{new_status}?")
       @notice = NoticeBuilder.new(new_status, new_status, "already_#{new_status}").as_json
     else
-      notification.contact.owner.send(action_method)
+      owner.send(action_method)
       @notice = NoticeBuilder.new(new_status, new_status, new_status).as_json
     end
   end
@@ -101,10 +102,14 @@ class WebClient::ActionHandler
   end
 
   def build_event
-    { triggered_by: "ff:#{caller}",
+    event = {
+      triggered_by: "ff:#{caller}",
       initiator: 'owner',
-      initiator_id: owner.mkey,
+      initiator_id: owner.mkey
+    }
+    event.merge!(
       target: 'notification',
-      target_id: notification.nkey }
+      target_id: notification.nkey) if notification
+    event
   end
 end